+++ libs/base/include/mrpt/utils/CFileOutputStream.h
@@ -97,15 +97,15 @@ namespace utils
 		/** Method for moving to a specified position in the streamed resource.
 		 *   See documentation of CStream::Seek
 		 */
-		size_t Seek(long Offset, CStream::TSeekOrigin Origin = sFromBeginning);
+		uint64_t Seek(long Offset, CStream::TSeekOrigin Origin = sFromBeginning);
 
 		/** Method for getting the total number of bytes writen to buffer.
 		 */
-		size_t getTotalBytesCount();
+		uint64_t getTotalBytesCount();
 
 		/** Method for getting the current cursor position, where 0 is the first byte and TotalBytesCount-1 the last one.
 		 */
-		size_t getPosition();
+		uint64_t getPosition();
 
 
 	}; // End of class def.
