+++ libs/base/include/mrpt/utils/CFileStream.h
@@ -111,23 +111,23 @@ namespace mrpt
 			/** Method for moving to a specified position in the streamed resource.
 			 *   See documentation of CStream::Seek
 			 */
-			size_t Seek(long Offset, CStream::TSeekOrigin Origin = sFromBeginning);
+			uint64_t Seek(long Offset, CStream::TSeekOrigin Origin = sFromBeginning);
 
 			/** Method for getting the total number of bytes writen to buffer.
 			 */
-			size_t getTotalBytesCount();
+			uint64_t getTotalBytesCount();
 
 			/** Method for getting the current cursor position, where 0 is the first byte and TotalBytesCount-1 the last one.
 			 */
-			size_t getPosition();
+			uint64_t getPosition();
 
 			/** The current Input cursor position, where 0 is the first byte.
 			 */
-			size_t getPositionI();
+			uint64_t getPositionI();
 
 			/** The current Input cursor position, where 0 is the first byte.
 			 */
-			size_t getPositionO();
+			uint64_t getPositionO();
 
 			/** Reads one string line from the file (until a new-line character)
 			  * \return true if a line has been read, false on EOF or error.
